requirements.txt
Flask==2.3.2
Pillow==8.3.2
openpyxl==3.0.9
